Clarifies some text, fixes a broken URL, adds Tails.
Rewords some of the discussion of security. Adds focus on open source
tools delivering on Kerckhoffs's principle.

The Onionshare github link was pointed to another tool prior to this
commit.

Adds the Tails liveCD as a convenient way to have a clean environment
free of malware & to get preloaded access to many of the tools already
listed on the page.

TODO: The Tails tool needs a screenshot/image